# docs/archive/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "World Weaver Memory Gateway API",
    "version": "1.0.0",
    "description": "MCP server providing access to a tripartite memory system (episodic, semantic, procedural) for Claude Code instances. Implements 17 tools for autobiographical events, knowledge graphs, and learned procedures with ACT-R activation, FSRS retrievability, and Hebbian learning.",
    "contact": {
      "name": "World Weaver Project",
      "url": "https://github.com/astoreyai/world-weaver"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local MCP server"
    }
  ],
  "paths": {
...
    "/tools/deprecate_skill": {
      "post": {
        "summary": "Mark a procedure as deprecated",
        "description": "Prevents the procedure from being retrieved in future searches",
        "operationId": "deprecate_skill",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "procedure_id": {
                    "type": "string",
                    "description": "Procedure UUID"
                  },
                  "reason": {
                    "type": "string",
                    "description": "Deprecation reason"
                  }
                },
                "required": [
                  "procedure_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "field": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}